requirements.txt
discord.py==2.5.2
feedparser==6.0.11
python-dotenv==1.1.0